module 01 verified

+++ Module_01/ex01/zombieHorde.cpp
@@ -6,7 +6,7 @@
 /*   By: adeimlin <adeimlin@student.42porto.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/04/15 15:40:48 by adeimlin          #+#    #+#             */
-/*   Updated: 2026/04/15 19:40:46 by adeimlin         ###   ########.fr       */
+/*   Updated: 2026/05/01 15:13:01 by adeimlin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,37 +23,7 @@ Zombie* zombieHorde(int N, std::string name)
 		return (NULL);
 	}
 	horde = new Zombie[N];
-	for (size_t i = 0; i < N; i++)
+	for (int i = 0; i < N; i++)
 		horde[i].setName(name);
 	return (horde);
 }
-
-// Zombie* zombieHorde(int N, std::string name)
-// {
-// 	Zombie *horde;
-
-// 	if (N <= 0)
-// 	{
-// 		std::cout << "A zombie horde cannot have negative zombies\n";
-// 		return (NULL);
-// 	}
-// 	horde = new Zombie[N];
-// 	for (size_t i = 0; i < N; i++)
-// 		horde[i] = Zombie(name);
-// 	return (horde);
-// }
-
-// Zombie* zombieHorde(int N, std::string name)
-// {
-// 	Zombie *horde;
-
-// 	if (N <= 0)
-// 	{
-// 		std::cout << "A zombie horde cannot have negative zombies\n";
-// 		return (NULL);
-// 	}
-// 	horde = (Zombie *) ::operator new (sizeof(Zombie) * N);
-// 	for (size_t i = 0; i < N; i++)
-// 		horde[i] = Zombie(name);
-// 	return (horde);
-// }
